Refactor agent package structure
