+++ agent/__init__.py
@@ -1,8 +1,13 @@
 from .chat import ChatSession
-from .team import TeamChatSession, send_to_junior, send_to_junior_async, set_team
-from .solo import SoloChatSession
+from .sessions.team import (
+    TeamChatSession,
+    send_to_junior,
+    send_to_junior_async,
+    set_team,
+)
+from .sessions.solo import SoloChatSession
 from .tools import execute_terminal, execute_terminal_async, set_vm
-from .utils import limit_chars
+from .utils.helpers import limit_chars
 from .vm import LinuxVM
 
 __all__ = [
